complete 09_sumAll
This is a little harder to read but I wanted to
commit to my inital solution, testing my ability
to use some of these unfamiliar array methods.

I am iterating through the two arguments ensuring
they are integers greater than 0, if they are not
the some method is able to break from the loop and
return a value, unlike the forEach method.

in the case where the conditional is true, it sets
the isInvalid flag as true and then returns error.

A cleaner solution would break this down and also
set min  and max rather than rewriting the same
code but in reverse